+++ tools/coreutils/printf/base.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+
+printf "\n"
+printf "\tx\n"