+++ CHANGELOG.md
@@ -27,7 +27,7 @@ and this project adheres to
 #### Scripts
 #### Tools
 - vim: add xxd subdir;
-- coreutils: add dd, od subdir;
+- coreutils: add dd, od, printf subdir;
 - redis: add stress subdir;
 - smartmontools: first commit;
 - dracut: add modules.d directory;